kaneko - [Bike] sizeの型変更 Integer ⇒ Double

+++ src/main/java/com/example/mmng_back/entity/Bike.java
@@ -33,7 +33,7 @@ public class Bike {
 	private String makerName;
 	
 	@Column(name = "size")
-	private Integer size;
+	private Double size;
 	
 	@Column(name = "wheel_base")
 	private Integer wheelBase;
